+++ web-local/src/routes/+layout.svelte
@@ -1,21 +1,61 @@
 <script lang="ts">
   import { RillTheme } from "@rilldata/web-common/layout";
+  import { featureFlags } from "@rilldata/web-common/features/feature-flags";
   import { initializeNodeStoreContexts } from "@rilldata/web-local/lib/application-state-stores/initialize-node-store-contexts";
-  import { runtime } from "@rilldata/web-common/runtime-client/runtime-store";
+  import { beforeNavigate } from "$app/navigation";
+  import { retainFeaturesFlags } from "@rilldata/web-common/features/feature-flags";
+  import { errorEventHandler } from "@rilldata/web-common/metrics/initMetrics";
+  import type { Query } from "@tanstack/query-core";
+  import { QueryClientProvider } from "@tanstack/svelte-query";
+  import type { AxiosError } from "axios";
+  import { runtimeServiceGetConfig } from "@rilldata/web-common/runtime-client/manual-clients";
+  import { queryClient } from "@rilldata/web-common/lib/svelte-query/globalQueryClient";
+  import WelcomePageRedirect from "@rilldata/web-common/features/welcome/WelcomePageRedirect.svelte";
+  import type { ApplicationBuildMetadata } from "@rilldata/web-common/layout/build-metadata";
+  import { initMetrics } from "@rilldata/web-common/metrics/initMetrics";
+  import { getContext, onMount } from "svelte";
+  import type { Writable } from "svelte/store";
+  import ResourceWatcher from "@rilldata/web-common/features/entity-management/ResourceWatcher.svelte";
 
-  export let data;
+  const appBuildMetaStore: Writable<ApplicationBuildMetadata> =
+    getContext("rill:app:metadata");
+
+  queryClient.getQueryCache().config.onError = (
+    error: AxiosError,
+    query: Query,
+  ) => errorEventHandler?.requestErrorEventHandler(error, query);
 
   /** This function will initialize the existing node stores and will connect them
    * to the Node server.
    */
   initializeNodeStoreContexts();
 
-  runtime.set({
-    host: data.host,
-    instanceId: data.instanceId,
+  beforeNavigate(retainFeaturesFlags);
+
+  export let data;
+
+  onMount(async () => {
+    const config = await runtimeServiceGetConfig();
+    await initMetrics(config);
+
+    featureFlags.set(false, "adminServer");
+    featureFlags.set(config.readonly, "readOnly");
+    // Disable AI when running e2e tests
+    featureFlags.set(!import.meta.env.VITE_PLAYWRIGHT_TEST, "ai");
+
+    appBuildMetaStore.set({
+      version: config.version,
+      commitHash: config.build_commit,
+    });
   });
 </script>
 
 <RillTheme>
-  <slot />
+  <QueryClientProvider client={queryClient}>
+    <WelcomePageRedirect>
+      <ResourceWatcher host={data.host} instanceId={data.instanceId}>
+        <slot />
+      </ResourceWatcher>
+    </WelcomePageRedirect>
+  </QueryClientProvider>
 </RillTheme>
